Move utils from shared file to separate files
For code readability, it is better to separate utility functions
into separate files.
